apps/api-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiomysql==0.2.0
PyMySQL==1.1.0 